Generalize HttpBindingResolver for protocols that don't have HTTP traits
